client_node/requirements.txt
flwr>=1.5.0
torch>=2.0.0
opacus>=1.4.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
httpx>=0.24.0